+++ sdk/healthbot/arm-healthbot/CHANGELOG.md
@@ -1,11 +1,24 @@
 # Release History
-
-## 2.0.1 (2022-04-26)
-
+    
+## 2.1.0 (2022-09-06)
+    
 **Features**
 
-  - Bug fix
-
+  - Added operation Bots.listSecrets
+  - Added operation Bots.regenerateApiJwtSecret
+  - Added Interface BotsListSecretsOptionalParams
+  - Added Interface BotsRegenerateApiJwtSecretOptionalParams
+  - Added Interface HealthBot
+  - Added Interface HealthBotKey
+  - Added Interface HealthBotKeysResponse
+  - Added Interface KeyVaultProperties
+  - Added Interface TrackedResource
+  - Added Type Alias BotsListSecretsResponse
+  - Added Type Alias BotsRegenerateApiJwtSecretResponse
+  - Interface HealthBotProperties has a new optional parameter keyVaultProperties
+  - Interface HealthBotUpdateParameters has a new optional parameter properties
+    
+    
 ## 2.0.0 (2022-01-13)
 
 The package of @azure/arm-healthbot is using our next generation design principles since version 2.0.0, which contains breaking changes.